refactor: 모임 운동 컨트롤러 분리
